Core FF: separate source files for IFormFactorPrism
