+++ Core/HardParticle/FormFactorPolyhedron.cpp
@@ -3,7 +3,7 @@
 //  BornAgain: simulate and fit scattering at grazing incidence
 //
 //! @file      Core/HardParticle/FormFactorPolyhedron.cpp
-//! @brief     Implements class FormFactorPolyhedron, FormFactorPrism, and auxiliary classes.
+//! @brief     Implements class FormFactorPolyhedron.
 //!
 //! @homepage  http://www.bornagainproject.org
 //! @license   GNU General Public License v3 or higher (see COPYING)
@@ -29,10 +29,6 @@ const double eps = 2e-16;
 
 double FormFactorPolyhedron::q_limit_series = 1e-2;
 int FormFactorPolyhedron::n_limit_series = 20;
-
-//**************************************************************************************************
-//  FormFactorPolyhedron implementation
-//**************************************************************************************************
 
 #ifdef POLYHEDRAL_DIAGNOSTIC
 void FormFactorPolyhedron::setLimits(double _q, int _n)
@@ -221,105 +217,4 @@ void FormFactorPolyhedron::assert_platonic() const
                       << " vs pyr_volume(avge)=" << pyramidal_volume << "\n";
             throw std::runtime_error("Deviant pyramidal volume in " + getName());
         }
-}
-
-//**************************************************************************************************
-//  FormFactorPolygonalPrism implementation
-//**************************************************************************************************
-
-FormFactorPolygonalPrism::FormFactorPolygonalPrism(const NodeMeta& meta,
-                                                   const std::vector<double>& PValues)
-    : IFormFactorBorn(meta, PValues)
-{
-}
-
-void FormFactorPolygonalPrism::setPrism(bool symmetry_Ci, const std::vector<kvector_t>& vertices)
-{
-    m_vertices.clear();
-    for (const kvector_t& vertex : vertices) {
-        m_vertices.push_back(vertex);
-        m_vertices.push_back(vertex + kvector_t{0, 0, height()});
-    }
-
-    try {
-        m_base = std::unique_ptr<PolyhedralFace>(new PolyhedralFace(vertices, symmetry_Ci));
-    } catch (std::invalid_argument& e) {
-        throw std::invalid_argument("Invalid parameterization of " + getName() + ": " + e.what());
-    } catch (std::logic_error& e) {
-        throw std::logic_error("Bug in " + getName() + ": " + e.what()
-                               + " [please report to the maintainers]");
-    } catch (std::exception& e) {
-        throw std::runtime_error("Unexpected exception in " + getName() + ": " + e.what()
-                                 + " [please report to the maintainers]");
-    }
-}
-
-double FormFactorPolygonalPrism::bottomZ(const IRotation& rotation) const
-{
-    return BottomZ(m_vertices, rotation);
-}
-
-double FormFactorPolygonalPrism::topZ(const IRotation& rotation) const
-{
-    return TopZ(m_vertices, rotation);
-}
-
-//! Returns the volume of this prism.
-double FormFactorPolygonalPrism::volume() const
-{
-    return height() * m_base->area();
-}
-
-//! Returns the form factor F(q) of this polyhedron, respecting the offset height/2.
-
-complex_t FormFactorPolygonalPrism::evaluate_for_q(cvector_t q) const
-{
-    try {
-#ifdef POLYHEDRAL_DIAGNOSTIC
-        diagnosis.maxOrder = 0;
-        diagnosis.nExpandedFaces = 0;
-#endif
-        cvector_t qxy(q.x(), q.y(), 0.);
-        return height() * exp_I(height() / 2 * q.z()) * MathFunctions::sinc(height() / 2 * q.z())
-               * m_base->ff_2D(qxy);
-    } catch (std::logic_error& e) {
-        throw std::logic_error("Bug in " + getName() + ": " + e.what()
-                               + " [please report to the maintainers]");
-    } catch (std::runtime_error& e) {
-        throw std::runtime_error("Numeric computation failed in " + getName() + ": " + e.what()
-                                 + " [please report to the maintainers]");
-    } catch (std::exception& e) {
-        throw std::runtime_error("Unexpected exception in " + getName() + ": " + e.what()
-                                 + " [please report to the maintainers]");
-    }
-}
-
-//**************************************************************************************************
-//  FormFactorPolygonalSurface implementation
-//**************************************************************************************************
-
-FormFactorPolygonalSurface::FormFactorPolygonalSurface(const NodeMeta& meta,
-                                                       const std::vector<double>& PValues)
-    : IFormFactorBorn(meta, PValues)
-{
-}
-
-complex_t FormFactorPolygonalSurface::evaluate_for_q(cvector_t q) const
-{
-    try {
-#ifdef POLYHEDRAL_DIAGNOSTIC
-        diagnosis.maxOrder = 0;
-        diagnosis.nExpandedFaces = 0;
-#endif
-        return m_base->ff(q, false);
-    } catch (std::logic_error& e) {
-        throw std::logic_error("Bug in " + getName() + ": " + e.what()
-                               + " [please report to the maintainers]");
-    } catch (std::runtime_error& e) {
-        throw std::runtime_error("Numeric computation failed in " + getName() + ": " + e.what()
-                                 + " [please report to the maintainers]");
-    } catch (std::exception& e) {
-        throw std::runtime_error("Unexpected exception in " + getName() + ": " + e.what()
-                                 + " [please report to the maintainers]");
-    }
 }
